Split standard and batched
